fix(prod): usar DB_PATH para persistencia SQLite + collectstatic en arranque y modularizar views
