+++ backend/Dockerfile
@@ -20,5 +20,5 @@ COPY . .
 # Exponer el puerto
 EXPOSE 8000
 
-# Ejecutar migraciones, seed data y arrancar con gunicorn
-CMD python manage.py migrate --noinput && gunicorn -w 3 reservia.wsgi:application --bind 0.0.0.0:8000
+# Ejecutar migraciones, seed data, collectstatic y arrancar con gunicorn
+CMD python manage.py migrate --noinput && python manage.py collectstatic --noinput && gunicorn -w 3 reservia.wsgi:application --bind 0.0.0.0:8000
